Create folder structure.
