+++ backend/Sources/Services/JokeAPI.swift
@@ -10,26 +10,6 @@ import Foundation
 import StORM
 import PostgresStORM
 
-struct AppConstants {
-    struct JokeKeys {
-        static let kSetup = "setup"
-        static let kPunchline = "punchline"
-        static let kVotes = "votes"
-        static let kID = "id"
-    }
-    
-    struct DatabaseConfig {
-        static let kHost = "localhost"
-        static let kUsername = "perfect"
-        static let kPassword = "password"
-        static let kDatabase = "nerdjokes"
-        static let kPort = 5432
-    }
-    
-    struct DatabaseTables {
-        static let kJoke = "joke"
-    }
-}
 
 class JokeAPI {
     private init() {}
